LayoutTests/imported/w3c/web-platform-tests/css/css-viewport/zoom/relative-units-expected.txt


PASS relative-units
FAIL relative-units 1 assert_approx_equals: rem in zoomed expected 40 +/- 1 but got 20
PASS relative-units 2
FAIL relative-units 3 assert_approx_equals: rlh in zoomed expected 40 +/- 1 but got 20
FAIL relative-units 4 assert_approx_equals: vh in outside expected 12 +/- 1 but got 6
FAIL relative-units 5 assert_approx_equals: vw in outside expected 16 +/- 1 but got 8
